backend/requirements.txt
flask
flask-cors
python-dotenv
requests
tavily-python
langchain
langchain-core
langchain-mistralai
rich